Made innodb_max_dirty_pages_pct my.cnf variable a double
Summary:
This change is to fix: http://bugs.mysql.com/62534

This makes innodb_max_dirty_pages_pct a double with min,default,max values
0.001, 75, 99.999.

This also makes innodb_max_dirty_pages_pct_lwm a double, as these sysvars
are inter-dependent.

Added more to the BUFFER POOL AND MEMORY section of SHOW INNODB STATUS:
Percent pages dirty: 8.939 --> this is all n_dirty_pages / used_pages
Percent all pages dirty: 0.195 --> this is n_dirty_pages / all-pages
Max dirty pages percent: 75.000 --> this is innodb_max_dirty_pages_pct

Also changed all of buf from 2 to 3 digits of precision (%.2f -> %.3f).

Test Plan: Jenkins

Reviewers: jtolmer, nizamordulu

Reviewed By: jtolmer

+++ storage/innobase/include/srv0srv.h
@@ -327,8 +327,8 @@ extern ulint	srv_win_file_flush_method;
 
 extern ulint	srv_max_n_open_files;
 
-extern ulong	srv_max_dirty_pages_pct;
-extern ulong	srv_max_dirty_pages_pct_lwm;
+extern double	srv_max_dirty_pages_pct;
+extern double	srv_max_dirty_pages_pct_lwm;
 
 extern ulong	srv_adaptive_flushing_lwm;
 extern ulong	srv_flushing_avg_loops;
@@ -355,7 +355,7 @@ extern my_bool			srv_stats_auto_recalc;
 extern ibool	srv_use_doublewrite_buf;
 extern ulong	srv_doublewrite_batch_size;
 
-extern ulong	srv_max_buf_pool_modified_pct;
+extern double	srv_max_buf_pool_modified_pct;
 extern ulong	srv_max_purge_lag;
 extern ulong	srv_max_purge_lag_delay;
